Moving isSupportedOceanNetwork to Web3 context
