+++ src/components/Header/Wallet/Network.tsx
@@ -5,23 +5,9 @@ import Tooltip from '@shared/atoms/Tooltip'
 import { useWeb3 } from '@context/Web3'
 import NetworkName from '@shared/NetworkName'
 import styles from './Network.module.css'
-import { getOceanConfig } from '@utils/ocean'
 
 export default function Network(): ReactElement {
-  const { networkId, isTestnet } = useWeb3()
-
-  const [isSupportedOceanNetwork, setIsSupportedOceanNetwork] =
-    useState<boolean>()
-
-  useEffect(() => {
-    // take network from user when present
-    const network = networkId || 1
-
-    // Check networkId against ocean.js ConfigHelper configs
-    // to figure out if network is supported.
-    const isSupportedOceanNetwork = Boolean(getOceanConfig(network))
-    setIsSupportedOceanNetwork(isSupportedOceanNetwork)
-  }, [networkId])
+  const { networkId, isTestnet, isSupportedOceanNetwork } = useWeb3()
 
   return networkId ? (
     <div className={styles.network}>
