+++ src/pages/publish.tsx
@@ -8,13 +8,15 @@ export default function PagePublish(): ReactElement {
   const { title, description } = content
 
   return (
-    <Page
-      title={title}
-      description={description}
-      uri={router.route}
-      noPageHeader
-    >
-      <Publish content={content} />
-    </Page>
+    <>
+      <Page
+        title={title}
+        description={description}
+        uri={router.route}
+        noPageHeader
+      >
+        <Publish content={content} />
+      </Page>
+    </>
   )
 }
